Add Migration

+++ Domain/Model/CharacterClassification.cs
@@ -2,7 +2,7 @@
 {
     public class CharacterClassification
     {
-        public short CharacterClassificationId { get; set; }
+        public byte CharacterClassificationId { get; set; }
         public string Classification { get; set; }
         public virtual ICollection<Character> Characters { get; set; }
     }
